feat: 캐릭터 키우기 기능 구현중

+++ Locket/domain/src/main/java/com/ssafy/locket/model/home/character/Toy.kt
@@ -6,5 +6,5 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 class Toy(
     val isAvailable: Boolean,
-    val remainingTimeMinutes: Long
+    val remainingTimeMinutes: Int
 ): BaseModel
